fix @require for directories
If @require is a directory, try to lookup style in package.json.
Fallback to index.css if no package.json

Currently, the css will be prefixed if a package.json file is found and
will not be prefixed if no package.json file is found. When loading a
dir without a package.json we should not prefix.